Added 5.3 upgrade option for Route::resource nested in a Route::group with a prefix
If a user had a 2 resource routes with same named controllers in 2 different Route::groups this breaks in L5.3. It was before mentioned that the solution is using names option on Route::resource, but that means we have to specify each method separately. If we simply wish to prefix entire route resource we can use 'as' option (this feature is not documented, but it is working in L5.3).