Cron entry example for schedule:run relative to project path
Ran into an issue where running the cron with a similar format to the existing example caused scheduled jobs to work unless they were scheduled commands because the call to `Application::formatCommandString()` within `Schedule::command()` appears to generate the command string assuming that the caller is already in the working directory of the project. eg: `php artisan some:command`.