Clarify the usage of the :count placeholder
There was no documentation of the default placeholder :count that you can use during localization. This caused confusion with the existing placeholder attribute documentation as users were thinking that they need to provide the value to trans_choice twice in order to display the exact value that caused the pluralization to take effect.