Improve view composer documentation
This adds a new subdirectory to the view composer example to be complementatry with the view creators below. Adds for a much nicer directory structure than the separate directories with the View prefix.

It also renames the service provider from `ComposerServiceProvider` to `ViewServiceProvider` to prevent confusion with the ambiguous "Composer" naming. Besides that it also doesn't hints that way that the service provider would be for composers only.