Move tip about 'htmlspecialchars' to first example
The docs currently display a tip and then immediately display a new section with the same text. This change displays the {tip} after the first example so that users can learn the information when it's relevant and keeps the text in the next section for users who skip the previous section.