Clarify Horizon balance default
At the moment it seemed as the simple balance strategy is the default strategy even if you leave it away but this isn't the case. Adding these words will make it clear that simple is only the default in the config file itself.

Ref.: https://github.com/laravel/horizon/issues/507