Eliminate (almost) all instances of "of course" and "as you might expect"
These phrases were overused, didn't add any value, and seemed
a bit condescending. People are reading the Laravel docs to
become educated. The docs shouldn't imply that everything
Laravel does is natural or a matter "of course", because that
could make people feel bad for having to learn how it works
instead of instinctively knowing it already.

fixes #4973